add api for adding fightstatistic

+++ src/dashbord/schemas/section_left_schemas.py
@@ -5,4 +5,4 @@ from decimal import Decimal
 
 class MetricsOutPut(BaseModel):
     metrics: str
-    score: Decimal
+    score: float
